refactor: 💡 rename comonent -> components
