+++ src/pages/contact/index.page.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useContactHandler } from '@/pages/contact/hooks/useContactHandler';
 import { ErrorState, initialInput, UserInput } from '@/pages/contact/type';
-import { ContactForm } from '@/pages/contact/comonents/ContactForm';
+import { ContactForm } from '@/pages/contact/components/ContactForm';
 
 export default function Contact() {
   const [userInput, setUserInput] = useState<UserInput>(initialInput);
